refactor: :hammer: refactor on modules
