+++ app/src/main/java/com/rodrigo/deeplarva/application/utils/Constants.kt
@@ -29,8 +29,6 @@ class Constants {
         const val BROADCAST_ACTION = "com.deeplarva.broadcast.NOTIFICATION"
 
         const val MESSAGE_SERVICE_STARTED = "Ejecutando conteo"
-        const val MESSAGE_SERVICE_DISCONNECTED = "Servicio sin comunicación"
-        const val MESSAGE_SERVICE_RUNNING = "Procesamiento en ejecución"
 
         const val SHARED_PREFERENCES_DEVICE_ID = "DEVICE-IDENTIFIER"
