+++ app/src/main/java/com/rodrigo/deeplarva/modules/prediction/FinalResult.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.ml
+package com.rodrigo.deeplarva.modules.prediction
 
 import android.graphics.Bitmap
