+++ app/src/main/java/com/rodrigo/deeplarva/routes/activity/PermissionsHandlerActivity.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.routes
+package com.rodrigo.deeplarva.routes.activity
 
 import android.content.Intent
 import android.content.pm.PackageManager
@@ -10,6 +10,7 @@ import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import com.rodrigo.deeplarva.R
 import com.rodrigo.deeplarva.application.utils.Constants
+import com.rodrigo.deeplarva.routes.PicturesActivity
 
 class PermissionsHandlerActivity: AppCompatActivity() {
