+++ app/src/main/java/com/rodrigo/deeplarva/routes/activity/PictureDetailActivity.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.routes
+package com.rodrigo.deeplarva.routes.activity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
