+++ app/src/main/java/com/rodrigo/deeplarva/routes/activity/SplashActivity.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.routes
+package com.rodrigo.deeplarva.routes.activity
 
 import android.content.Intent
 import android.content.pm.PackageManager
@@ -8,6 +8,7 @@ import androidx.core.content.ContextCompat
 import com.rodrigo.deeplarva.R
 import com.rodrigo.deeplarva.application.usecases.UseCaseRegisterDeviceId
 import com.rodrigo.deeplarva.application.utils.Constants
+import com.rodrigo.deeplarva.routes.PicturesActivity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.delay
