+++ app/src/main/java/com/rodrigo/deeplarva/ui/tasks/BackgroundTaskPredict.kt
@@ -7,7 +7,7 @@ import android.os.Environment
 import androidx.annotation.RequiresApi
 import com.rodrigo.deeplarva.application.utils.Constants
 import com.rodrigo.deeplarva.domain.entity.Picture
-import com.rodrigo.deeplarva.ml.Detect640x640
+import com.rodrigo.deeplarva.modules.prediction.Detect640x640
 import com.rodrigo.deeplarva.utils.BitmapUtils
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
